added swal, delete test in progress

+++ client/src/Home/components/ListPasswords.js
@@ -3,11 +3,35 @@ import db from "../../config/firestore";
 import useFetcher from '../../hooks/useFetcher'
 import { useDispatch } from 'react-redux'
 import { deletePassword } from '../../store/actions'
+import Swal from 'sweetalert2'
 export function ListPasswords(props) {
   const { data } = props
   const dispatch = useDispatch()
   const handleDelete = id => {
-    dispatch(deletePassword(id))
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+        dispatch(deletePassword(id))
+        Swal.fire(
+          'Deleted!',
+          'Your password has been deleted.',
+          'success'
+        )
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Your password is safe :)',
+          'error'
+        )
+      }
+    })
+    
   };
 
   return (
@@ -29,7 +53,7 @@ export function ListPasswords(props) {
                 <span className="text-blue-500 block mb-2">
                   password: {password.password}{" "}
                 </span>
-                <button
+                <button data-testid="delete-button"
                   onClick={() => handleDelete(password.id)}
                   className="px-2 py-2 bg-red-400 text-white rounded"
                 >
